flag and unflag extracted and tested
